LEB128 - Change to big integer as it allows 128 bits
The earlier prototype worked with long, was a wip, fixed to use BigInteger,
add smoke tests, java-client can still use long as reaching collection
of greater than Long.MAX_VALUE has less chances, probably need to be
changed later on.
BREAKS SCALA!!

+++ java-client/src/main/java/com/couchbase/client/java/AsyncCollection.java
@@ -31,6 +31,7 @@ import com.couchbase.client.java.env.ClusterEnvironment;
 import com.couchbase.client.java.kv.*;
 import io.opentracing.Scope;
 
+import java.math.BigInteger;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
@@ -110,7 +111,7 @@ public class AsyncCollection {
     this.coreContext = core.context();
     this.environment = environment;
     this.bucket = bucket;
-    this.collectionId = UnsignedLEB128.encode(id);
+    this.collectionId = UnsignedLEB128.encode(BigInteger.valueOf(id));
     this.asyncBinaryCollection = new AsyncBinaryCollection(core, environment, bucket, collectionId);
   }
